unsupervised/requirements.txt
numpy
scipy
scikit-learn
numba
tqdm
pynndescent
matplotlib
datashader
holoviews
tensorflow
umap-learn
hdbscan
xgboost
librosa
joblib
xgboost
